requirements.txt
python-escpos
flask
